step1 - refactor
