+++ src/test/java/nextstep/member/acceptance/MemberSteps.java
@@ -75,16 +75,4 @@ public class MemberSteps {
                 .statusCode(HttpStatus.OK.value())
                 .extract();
     }
-
-    public static String 로그인_요청_후_토큰발급(String email, String password) {
-        return RestAssured.given().log().all()
-                .body(new TokenRequest(email, password))
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/login/token")
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value())
-                .extract()
-                .jsonPath()
-                .getString("accessToken");
-    }
 }
